build: Factor out and share finder-making code
* Move the code that checks if the concrete Landmarks finders need to be
  regenerated.
* Use this code in the profile script, negating the need for the
  --pre-process mode of the build script.
* Remove the pre-processing mode of the build script.
